drm/panel: simple: Adjust n116bge mode & rate for veyrons
This should get us a close to 60 Hz mode (60.00067..)
from the rk3288 available clock of 74.25 Mhz.

This has htotal of 1543 and vtotal 802. According
to a datasheet for the N116BGE it should accept vtotal
from 1506 to 1716, but in my testing it is stable
from 1520 onwards. There would be closer to 60Hz
modes at 1511*819 (loses sync) and at 1524*812 (works),
but i thought the latter was a bit too close to the edge,
so this is the next best mode.

Signed-off-by: Urja Rannikko <urjaman@gmail.com>